requirements.txt
Flask>=3.0,<4
Pillow>=10,<12
cryptography>=42,<45
waitress>=3,<4
python-dotenv>=1,<2
requests>=2,<3
